Apollo configuration improved
Apollo now should support SSR/ISR. This is not the best implementation, but it is the best I could find and build.

Apollo should be used inside the getServerSideProps/getStaticProps and inside the component (in different ways, obviously). The cache is shared between the SSR/SSG/ISR and the client side, so there shouldn't be duplicated requests to the backend, this means that no code inside the components should be rewritten, only inside the SSR/SSG/ISR functions and only if they are required for building the damn thing